Fix crash where you clear logs and immediately go back to view an empty list.

+++ aclient/src/main/java/dt/call/aclient/sqlite/SQLiteDb.java
@@ -185,19 +185,23 @@ public class SQLiteDb extends SQLiteOpenHelper
 
 	public ArrayList<DBLog> getLogs()
 	{
+		ArrayList<DBLog> result = new ArrayList<DBLog>();
+
 		//clear out old logs. unlikely to be viewed and just slow down the log ui viewer
 		final String idquery = "select " + colId + " from logs order by " + colId + " desc limit 1";
 		Cursor idcursor = appdb.rawQuery(idquery, null);
-		idcursor.moveToFirst();
-		long mostRecent = idcursor.getLong(0);
-		Long tooOld = mostRecent - Const.LOG_LIMIT;
-		if(tooOld > 0)
+		if(idcursor.moveToFirst())
 		{
-			appdb.delete(tableLogs, colId+"<=?", new String[]{tooOld.toString()});
+			long mostRecent = idcursor.getLong(0);
+			Long tooOld = mostRecent - Const.LOG_LIMIT;
+			if(tooOld > 0)
+			{
+				appdb.delete(tableLogs, colId+"<=?", new String[]{tooOld.toString()});
+			}
 		}
+		idcursor.close();
 
 		//now get the more recent logs
-		ArrayList<DBLog> result = new ArrayList<DBLog>();
 		final String query = "select * from logs order by " + colLogTs + " desc";
 		Cursor cursor = appdb.rawQuery(query, null);
 		cursor.moveToFirst();
